Allow purging to be disabled

+++ src/io/github/dead_i/bungeeweb/BungeeWeb.java
@@ -169,9 +169,10 @@ public class BungeeWeb extends Plugin {
 
     public void setupPurging(String type) {
         int days = getConfig().getInt("server." + type + "days");
-        int purge = getConfig().getInt("server.purge");
-        purge = (purge < 1 ? 10 : purge);
-        getProxy().getScheduler().schedule(this, new PurgeScheduler("stats", (days < 1 ? 30 : days)), purge, purge, TimeUnit.MINUTES);
+        int purge = getConfig().getInt("server.purge", 10);
+        if (purge > 0) {
+            getProxy().getScheduler().schedule(this, new PurgeScheduler("stats", (days < 1 ? 30 : days)), purge, purge, TimeUnit.MINUTES);
+        }
     }
 
     public static Configuration getConfig() {
